+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Armor and shield blueprints not working from the blueprint book
 - Repair kits repairing items with lower tier materials
 - Block placer traits not consuming durability or playing sound effects
+- Pressing "Cycle Back" when viewing material info causing a crash in some cases
 
 ## [2.0.0] - 2020-07-26
 Major overhaul of model system (again)